+++ app/lib/init.dart
@@ -8,6 +8,7 @@ import 'package:flutter/foundation.dart';
 import 'package:flutter/widgets.dart';
 import 'package:flutter_displaymode/flutter_displaymode.dart';
 import 'package:localsend_app/pages/home_page.dart';
+import 'package:localsend_app/pages/home_page_controller.dart';
 import 'package:localsend_app/provider/animation_provider.dart';
 import 'package:localsend_app/provider/app_arguments_provider.dart';
 import 'package:localsend_app/provider/device_info_provider.dart';
@@ -33,6 +34,7 @@ import 'package:localsend_app/util/native/cache_helper.dart';
 import 'package:localsend_app/util/native/context_menu_helper.dart';
 import 'package:localsend_app/util/native/cross_file_converters.dart';
 import 'package:localsend_app/util/native/device_info_helper.dart';
+import 'package:localsend_app/util/native/open_file_receiver.dart';
 import 'package:localsend_app/util/native/platform_check.dart';
 import 'package:localsend_app/util/native/tray_helper.dart';
 import 'package:localsend_app/util/ui/dynamic_colors.dart';
@@ -127,7 +129,7 @@ Future<RefenaContainer> preInit(List<String> args) async {
 StreamSubscription? _sharedMediaSubscription;
 
 /// Will be called when home page has been initialized
-Future<void> postInit(BuildContext context, Ref ref, bool appStart, void Function(int) goToPage) async {
+Future<void> postInit(BuildContext context, Ref ref, bool appStart) async {
   await updateSystemOverlayStyle(context);
 
   if (checkPlatform([TargetPlatform.android])) {
@@ -154,10 +156,26 @@ Future<void> postInit(BuildContext context, Ref ref, bool appStart, void Functio
 
   if (appStart) {
     final args = ref.read(appArgumentsProvider);
-    await ref.global.dispatchAsync(_HandleAppStartArgumentsAction(
-      args: args,
-      goToPage: goToPage,
-    ));
+
+    if (defaultTargetPlatform == TargetPlatform.macOS) {
+      final files = await getOpenedFiles();
+      if (files.isNotEmpty) {
+        await ref.global.dispatchAsync(_HandleAppStartArgumentsAction(
+          args: files,
+        ));
+      }
+
+      // handle future dropped files
+      getOpenedFilesStream().listen((files) {
+        ref.global.dispatchAsync(_HandleAppStartArgumentsAction(
+          args: files,
+        ));
+      });
+    } else {
+      await ref.global.dispatchAsync(_HandleAppStartArgumentsAction(
+        args: args,
+      ));
+    }
   }
 
   bool hasInitialShare = false;
@@ -172,7 +190,6 @@ Future<void> postInit(BuildContext context, Ref ref, bool appStart, void Functio
         // ignore: unawaited_futures
         ref.global.dispatchAsync(_HandleShareIntentAction(
           payload: initialSharedPayload,
-          goToPage: goToPage,
         ));
       }
     }
@@ -181,7 +198,6 @@ Future<void> postInit(BuildContext context, Ref ref, bool appStart, void Functio
     _sharedMediaSubscription = shareHandler.sharedMediaStream.listen((SharedMedia payload) {
       ref.global.dispatchAsync(_HandleShareIntentAction(
         payload: payload,
-        goToPage: goToPage,
       ));
     });
   }
@@ -202,11 +218,9 @@ Future<void> postInit(BuildContext context, Ref ref, bool appStart, void Functio
 
 class _HandleShareIntentAction extends AsyncGlobalAction {
   final SharedMedia payload;
-  final void Function(int) goToPage;
 
   _HandleShareIntentAction({
     required this.payload,
-    required this.goToPage,
   });
 
   @override
@@ -220,24 +234,22 @@ class _HandleShareIntentAction extends AsyncGlobalAction {
           converter: CrossFileConverters.convertSharedAttachment,
         ));
 
-    goToPage(HomeTab.send.index);
+    ref.redux(homePageControllerProvider).dispatch(ChangeTabAction(HomeTab.send));
   }
 }
 
 class _HandleAppStartArgumentsAction extends AsyncGlobalAction {
   final List<String> args;
-  final void Function(int) goToPage;
 
   _HandleAppStartArgumentsAction({
     required this.args,
-    required this.goToPage,
   });
 
   @override
   Future<void> reduce() async {
     final filesAdded = await ref.redux(selectedSendingFilesProvider).dispatchAsyncTakeResult(LoadSelectionFromArgsAction(args));
     if (filesAdded) {
-      goToPage(HomeTab.send.index);
+      ref.redux(homePageControllerProvider).dispatch(ChangeTabAction(HomeTab.send));
     }
   }
 }