+++ app/lib/pages/tabs/receive_tab.dart
@@ -1,10 +1,10 @@
 import 'package:flutter/material.dart';
 import 'package:localsend_app/gen/strings.g.dart';
 import 'package:localsend_app/pages/home_page.dart';
+import 'package:localsend_app/pages/home_page_controller.dart';
 import 'package:localsend_app/pages/receive_history_page.dart';
 import 'package:localsend_app/pages/tabs/receive_tab_vm.dart';
 import 'package:localsend_app/provider/animation_provider.dart';
-import 'package:localsend_app/provider/ui/home_tab_provider.dart';
 import 'package:localsend_app/util/ip_helper.dart';
 import 'package:localsend_app/widget/animations/initial_fade_transition.dart';
 import 'package:localsend_app/widget/custom_icon_button.dart';
@@ -42,7 +42,7 @@ class ReceiveTab extends StatelessWidget {
                               delay: const Duration(milliseconds: 200),
                               child: Consumer(builder: (context, ref) {
                                 final animations = ref.watch(animationProvider);
-                                final activeTab = ref.watch(homeTabProvider);
+                                final activeTab = ref.watch(homePageControllerProvider.select((state) => state.currentTab));
                                 return RotatingWidget(
                                   duration: const Duration(seconds: 15),
                                   spinning: vm.serverState != null && animations && activeTab == HomeTab.receive,